Adds minor refinements to LinuxParser routines

+++ include/linux_parser.h
@@ -19,7 +19,7 @@ const std::string kOSPath{"/etc/os-release"};
 const std::string kPasswordPath{"/etc/passwd"};
 
 // System
-float MemoryUtilization();
+double MemoryUtilization();
 long UpTime();
 std::vector<int> Pids();
 int TotalProcesses();
@@ -58,7 +58,7 @@ std::string Command(int pid);
 std::string Ram(int pid);
 std::string Uid(int pid);
 std::string User(int pid);
-long int UpTime(int pid);
+long UpTime(int pid);
 };  // namespace LinuxParser
 
 #endif
